@@ -1,7 +1,7 @@
-package vas.com.githubrepositories
+package com.example.githubrepositories
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 
 class MainActivity : AppCompatActivity() {
 
